moved menu template into a separate file
